Change skill ratings back to between 0 and 1.

+++ src/PersonProfile.java
@@ -2,13 +2,13 @@ import java.util.HashSet;
 
 public class PersonProfile {
     public int id;
-    public int[] skills;
+    public double[] skills;
     public HashSet<Integer> preferredPartners;
     public HashSet<Integer> silverBullets;
 
     public PersonProfile(int id, int numSkills) {
         this.id = id;
-        skills = new int[numSkills];
+        skills = new double[numSkills];
         preferredPartners = new HashSet<>();
         silverBullets = new HashSet<>();
     }
